requirements.txt
aiogram==2.11.2
aiohttp==3.7.3
aiohttp-healthcheck==1.3.1
aiojobs==0.2.2
aiomysql==0.0.21
aioredis==1.3.1
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
cryptography==3.3.1
hiredis==1.1.0
idna==2.10
loguru==0.5.1
multidict==4.7.6
peewee==3.14.0
peewee-async==0.7.1
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.20
PyMySQL==0.9.3
python-decouple==3.3
pytz==2020.4
rsa==4.6
six==1.15.0
Telethon==1.18.2
typing-extensions==3.7.4.3
win32-setctime==1.0.3
yarl==1.6.3
